minor shuffle
